Sample within square of pixel
The previous sampling could potentially result in us sampling
very close to if not directly on the next pixel, making the lines
on the image look very uneven and improper. This change
limits the randomised sampling to within 0.9 of the pixel's center,
so we still can get lots of coverage with our samples, but won't
risk going too far from the center.